backend/requirements.txt
# Python environment management
python-dotenv==0.19.2

# Binance API library
python-binance==1.0.15

# Machine learning libraries
tensorflow==2.8.0
scikit-learn==0.24.2

# Data manipulation
pandas==1.3.5

# Web framework
flask==2.0.3

# Other dependencies
requests==2.26.0